backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Redis / Queue
redis==5.2.1
celery==5.4.0

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
httpx==0.28.1

# Dev
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
